+++ curvlinops/__init__.py
@@ -5,6 +5,7 @@ from curvlinops.diagonal.hutchinson import hutchinson_diag
 from curvlinops.ekfac import EKFACLinearOperator
 from curvlinops.fisher import FisherMCLinearOperator
 from curvlinops.ggn import GGNLinearOperator
+from curvlinops.ggn_diagonal import GGNDiagonalLinearOperator
 from curvlinops.gradient_moments import EFLinearOperator
 from curvlinops.hessian import HessianLinearOperator
 from curvlinops.inverse import (
@@ -31,6 +32,7 @@ __all__ = [
     # linear operators
     "HessianLinearOperator",
     "GGNLinearOperator",
+    "GGNDiagonalLinearOperator",
     "EFLinearOperator",
     "FisherMCLinearOperator",
     "KFACLinearOperator",